# docs/swagger.yaml
basePath: /
definitions:
...
  model.AuditLog:
    properties:
      action:
        type: string
      created_at:
        type: string
      duration:
        description: 毫秒
        type: integer
      error:
        type: string
      id:
        type: integer
      ip:
        type: string
      method:
        type: string
      path:
        type: string
      request:
        type: string
      resource:
        type: string
      resource_id:
        type: integer
      response:
        type: string
      status:
        type: integer
      user:
        allOf:
        - $ref: '#/definitions/model.User'
        description: 关联
      user_agent:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  model.ConfigValue:
    additionalProperties: true
    type: object
...
  model.PaginationResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  model.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  model.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      updated_at:
        type: string
    required:
    - name
    type: object
...
  model.SystemConfig:
    properties:
      category:
        type: string
      created_at:
        type: string
      default_value:
        $ref: '#/definitions/model.ConfigValue'
      description:
        type: string
      display_name:
        type: string
      id:
        type: integer
      is_editable:
        type: boolean
      is_public:
        type: boolean
      key:
        type: string
      type:
        type: string
      updated_at:
        type: string
      validation:
        type: string
      value:
        $ref: '#/definitions/model.ConfigValue'
    type: object
  model.SystemConfigUpdateRequest:
    properties:
      value:
        $ref: '#/definitions/model.ConfigValue'
    required:
    - value
    type: object
...
  model.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      department:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      phone:
        type: string
      real_name:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/model.Role'
        description: 关联
      role_id:
        type: integer
      status:
        description: 1-正常，0-禁用
        type: integer
      updated_at:
        type: string
      username:
        type: string
    required:
    - email
    - real_name
    - role_id
    - username
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@vulnark.com
    name: VulnArk Team
    url: https://vulnark.com
  description: VulnArk漏洞管理系统API文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://vulnark.com/terms
  title: VulnArk API
  version: 1.0.0
paths:
...
  /api/v1/public/configs:
    get:
      description: 获取可公开访问的系统配置
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.SystemConfig'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 获取公开配置列表
      tags:
      - 系统管理
...
  /api/v1/system/audit-logs:
    get:
      description: 获取审计日志列表（分页）
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      - description: 用户ID
        in: query
        name: user_id
        type: string
      - description: 用户名
        in: query
        name: username
        type: string
      - description: 操作动作
        in: query
        name: action
        type: string
      - description: 资源类型
        in: query
        name: resource
        type: string
      - description: 请求方法
        in: query
        name: method
        type: string
      - description: 响应状态
        in: query
        name: status
        type: string
      - description: IP地址
        in: query
        name: ip
        type: string
      - description: 开始日期
        in: query
        name: start_date
        type: string
      - description: 结束日期
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.PaginationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取审计日志列表
      tags:
      - 系统管理
  /api/v1/system/audit-logs/{id}:
    get:
      description: 根据ID获取审计日志详情
      parameters:
      - description: 日志ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AuditLog'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取审计日志详情
      tags:
      - 系统管理
  /api/v1/system/audit-logs/cleanup:
    post:
      description: 清理指定天数之前的审计日志
      parameters:
      - description: 保留天数
        in: query
        name: days
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 清理审计日志
      tags:
      - 系统管理
  /api/v1/system/audit-logs/stats:
    get:
      description: 获取审计日志统计信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取审计日志统计
      tags:
      - 系统管理
  /api/v1/system/configs:
    get:
      description: 获取所有系统配置
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.SystemConfig'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取系统配置列表
      tags:
      - 系统管理
  /api/v1/system/configs/{key}:
    get:
      description: 根据键获取系统配置详情
      parameters:
      - description: 配置键
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.SystemConfig'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取系统配置详情
      tags:
      - 系统管理
    put:
      consumes:
      - application/json
      description: 更新系统配置值
      parameters:
      - description: 配置键
        in: path
        name: key
        required: true
        type: string
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.SystemConfigUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新系统配置
      tags:
      - 系统管理
  /api/v1/system/configs/category/{category}:
    get:
      description: 根据分类获取系统配置列表
      parameters:
      - description: 配置分类
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.SystemConfig'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 根据分类获取配置列表
      tags:
      - 系统管理
...
securityDefinitions:
  ApiKeyAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"